requirements.txt
cryptography==2.1.4
py-dmidecode==0.1.0
